Moved SelectionCardAdapter
